build(ps2): reject implicit float to double promotion (P3)
The EE FPU is single precision, so an implicit promotion to double runs in
soft-float. PS2 modules now compile with -Werror=double-promotion; FTicker
converts to its double clock explicitly. Budgets.md records the P3 sizes:
ThirdPerson is unchanged (section GC drops the unused math), TestPAL grows
by 92 KB of text and passes 35 tests in PCSX2.

+++ Engine/Source/Runtime/Core/Private/Containers/Ticker.cpp
@@ -11,7 +11,7 @@ FTicker& FTicker::GetCoreTicker()
 FDelegateHandle FTicker::AddTicker(const FTickerDelegate& InDelegate, float InDelay)
 {
 	FElement& Element = Elements.AddDefaulted_GetRef();
-	Element.FireTime = CurrentTime + InDelay;
+	Element.FireTime = CurrentTime + double(InDelay);
 	Element.DelayTime = InDelay;
 	Element.Delegate = InDelegate;
 	return InDelegate.GetHandle();
@@ -44,7 +44,7 @@ void FTicker::RemoveTicker(FDelegateHandle Handle)
 
 void FTicker::Tick(float DeltaTime)
 {
-	CurrentTime += DeltaTime;
+	CurrentTime += double(DeltaTime);
 	bInTick = true;
 
 	// Delegates added during this Tick fire next frame.
@@ -66,7 +66,7 @@ void FTicker::Tick(float DeltaTime)
 		}
 		else
 		{
-			Elements[Index].FireTime = CurrentTime + Elements[Index].DelayTime;
+			Elements[Index].FireTime = CurrentTime + double(Elements[Index].DelayTime);
 		}
 	}
